+++ internal/testdata/metric.go
@@ -279,10 +279,10 @@ func initMetric(m pdata.Metric, name string, ty pdata.MetricDataType) {
 	case pdata.MetricDataTypeSum:
 		sum := m.Sum()
 		sum.SetIsMonotonic(true)
-		sum.SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
+		sum.SetAggregationTemporality(pdata.MetricAggregationTemporalityCumulative)
 	case pdata.MetricDataTypeHistogram:
 		histo := m.Histogram()
-		histo.SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
+		histo.SetAggregationTemporality(pdata.MetricAggregationTemporalityCumulative)
 	}
 }